pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hack-balatro"
version = "0.1.0"
description = "Reproducible Balatro RL rebuild pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pyyaml>=6.0",
  "gymnasium>=0.29",
  "torch>=2.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.5.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
